Platform/Sgi: Add ProductId lookup values for RD-N2-Cfg2 Platform
The RD-N2-Cfg2 platform is a quad-chip variant of the RD-N2 platform
but with reduced core count, that is, each instance of the RD-N2 chip
has four cores. This platform shares the same product and config ID
as the RD-N2 platform.

As a preparatory step towards adding support for RD-N2-Cfg2 Platfrom,
add the Product ID lookup values for GetProductID API.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Signed-off-by: Pranav Madhu <pranav.madhu@arm.com>