Silicon/Qemu/Sbsa: Rectify SSDT length calculations
The SSDT table so far produced incorrect length fields in both the SSDT
table header and the scope. The scope length has so far been set for
the maximum size, and the surrounding table length has been clamped down
to a smaller size.

The SetPkgLength function should be passed the actual size of its
payload, and then needs to increase by the amount of bytes needed to
represent the length field.  In addition the TableSize adjustment
correction needs to consider the length of the scope opcode.

Signed-off-by: Patrick Wildt <patrick@blueri.se>
Acked-by: Ard Biesheuvel <ardb@kernel.org>