Platform/ARM: Add BootDiscoveryPolicy
This commit adds code which check BootDiscoveryPolicy variable and
calls Boot Policy Manager Protocol to connect device specified by
the variable. To enable that mechanism for platform
EfiMdeModulePkgTokenSpaceGuid.PcdBootDiscoveryPolicy PCD must be
added to DSC file.
For the edk2 repository, commit cae735f61328 ("ArmPkg: Enable boot
discovery policy for ARM package.") adds a mechanism to check for
the boot discovery policy. When the following are available:
- gEfiBootManagerPolicyProtocolGuid protocol
- BOOT_DISCOVERY_POLICY_VAR variable
this allows to choose whether devices need to be connected during
boot.

During boot on ARM platforms, all devices are connected if there
are no valid boot options. If the UEFI shell is added as a valid
boot option, some devices won't be connected. The BootDiscoveryPolicy
will allow to manually configure this.

For all ARM platforms, add:
- BootManagerPolicyDxe module to Bds phase
- BootDiscoveryPolicyUiLib to UiApp module
- PcdBootDiscoveryPolicy Pcd

Signed-off-by: Pierre Gondois <Pierre.Gondois@arm.com>
Reviewed-by: Ard Biesheuvel <ardb@kernel.org>