Platform/ARM: FVP: set IORT RC address limit for FVP RevC
When booting Linux on FVP RevC, Linux reports the following error:

ACPI: IORT: [Firmware Bug]: Root complex missing memory address limit

This issue is due to the RC Address Limit not being configured in the
configuration manager for the platform.

Set the value to 64 as the RevC FVP can use the full width, fixing the
error.

Signed-off-by: Edward Pickup <edward.pickup@arm.com>
Reviewed-by: Ard Biesheuvel <ardb@kernel.org>