Platform/Sgi: Remove inaccurate comment in dsc file
The base address of NV storage are different on reference design
platforms after the introduction of the RD-N2 platform. So the comment
for the NV storage sizes for variables is corrected as it is now
inaccurate. The base address of NV storage is correctly listed in the
SgiMemoryMap.dsc.inc and SgiMemoryMap2.dsc.inc files.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Reviewed-by: Thomas Abraham <thomas.abraham@arm.com>