AmpereAltraPkg: Remove MemoryAllocationLib in ArmPlatformLib
PrePeiCore's Arm module just added a requirement to run all library
constructors that PrePeiCore uses. One of them is included in the
constructor of the ExtractGuidedSectionLib. This constructor will build
a HOB that is not supported "at this time" and causes an error.
We don't need to use this constructor because PrePeiCore doesn't
really need this library.

This is a workaround to avoid PrePeiCore module calling
ExtractGuidedSectionLib's constructor.

Signed-off-by: Nhi Pham <nhi@os.amperecomputing.com>