Revert "Platform/Sgi: Fix missing FdtLib instance issue"
This reverts commit eb3950b842d3f335671600fdfa9b58c482886c79.

This commit referenced the change (9ad168c9e0 StandaloneMmPkg:
Include libfdt in the StMM) in edk2 as the dependency to add FdtLib
instance. However, as this change is not yet upstreamed, it was
incorrect to introduce this dependency and refer to the change in
the commit message. So reverting this patch.

Signed-off-by: Vivek Gautam <vivek.gautam@arm.com>
Reviewed-by: Sami Mujawar <sami.mujawar@arm.com>