Platform/ARM: Fix BootMonFS device path
The NOR Flash driver was recently moved from the Tianocore\edk2 repository
to the Tianocore\edk2-platforms repository at the following location:
Platform\ARM\Drivers\NorFlashDxe\NorFlashDxe.inf

As part of this move the FILE_GUID for the NorFlashDxe.inf at the new
location was also updated from: 93E34C7E-B50E-11DF-9223-2443DFD72085
to: DE6AE758-D662-4E17-A97C-4C5964DA4C41

Correspondingly, the device paths for the BootMonFs also requires to be
updated in gArmBootMonFsTokenSpaceGuid.PcdBootMonFsSupportedDevicePaths
so that the file system is correctly mounted.

Therefore, update the BootMonFS device path for the platforms that
utilise it.

Reported-by: Sudeep Holla <Sudeep.Holla@arm.com>
Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>
Tested-by: Sudeep Holla <Sudeep.Holla@arm.com>