Platform/BeagleBoard: Drop PeCoff protocol
BeagleBoard is the only user of the so-called 'PE/COFF protocol', which
just exposes the PE/COFF loader library API via a protocol, presumably
to avoid duplicating this code in the PrePi SEC component as well as the
DXE core.

This is a rather questionable practice, and it would be better to drop
this code so we can remove it from the EDK2 main repo as well.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Reviewed-by: Leif Lindholm <quic_llindhol@quicinc.com>