Silicon/Marvell: Drop use of NONSECURE ARM memory region type
The ARM memory region types with the NONSECURE qualification should only
be used in places where this makes a difference, i.e., in the secure
world. The Marvell firmware generally runs at EL2, which is non-secure
only (in ARM v8.0) so the NONSECURE marking is unnecessary here, and
should be dropped.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Reviewed-by: Leif Lindholm <quic_llindhol@quicinc.com>