SimicsOpenBoardPkg: Use another SmmAccess Driver
Because of the similar reason I mentioned in last commit, the
SmmAccess2Dxe.inf driver should be deleted and the replacement
will avoid hard-code and use gEfiSmmSmramMemoryGuid Hob to get
Smm Range information.

This can fix an existing bug. In pei phase, some module may use some
SMM range, and mark this range as allocated in gEfiSmmSmramMemoryGuid
Hob. In DXE phase, when loading SMM core, we should avoid using the
allocated SMM range. However, because the GetCapabilities by this
driver return a hard-code SMM range information, the allocated buffer
is treated as free buffer and used to load SMM core.

Cc: Nate DeSimone <nathaniel.l.desimone@intel.com>
Reviewed-by: Ray Ni <ray.ni@intel.com>
Signed-off-by: Zhiguang Liu <zhiguang.liu@intel.com>