AmpereAltraPkg: Add retry logic when PCIe card is present
Add PCIe link retry logic. If the card is detected to be present, allow
a retry for 3 times.

When the link is down, check if the LTSSMEN bit is cleared.  Based on
Altra PCIe IP, the LTSSMEN bit is cleared from 1 to 0 if there is a
surprising link down.  This indicates that a card is present.

Signed-off-by: Minh Nguyen <minhnguyen1@os.amperecomputing.com>
Reviewed-by: Nhi Pham <nhi@os.amperecomputing.com>