Platform/Arm: Disable memmap platform timers for JunoR0
Juno includes a system-level wakeup timer which is an implementation
of the ARM Generic Timer architecture.
The CNTCTL frame contains some registers that are accessible using a
Non-Secure access. Juno R0 incorrectly limits these accesses to Secure
access only. This issue is documented in the Juno Errata
832219: APB port security breaks SBSA compliance
https://developer.arm.com/documentation/epm008857/latest

This results in a crash when the OS tries to access these registers.
Therefore, disable memory mapped platform timers for Juno R0.

Cc: Ard Biesheuvel <ardb+tianocore@kernel.org>
Cc: Thomas Abraham <thomas.abraham@arm.com>
Cc: Sudeep Holla <Sudeep.Holla@arm.com>

Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>
Reported-by: Sudeep Holla <Sudeep.Holla@arm.com>
Tested-by: Sudeep Holla <Sudeep.Holla@arm.com>
Reviewed-by: Pierre Gondois <pierre.gondois@arm.com>