edk2-platform: Add Readme under Platform\AMD
For AMD edk2 opensource project and the needs
of opensource AMD platform common modules those can
be leverage across AMD business units, AMD server,
client and S3 (silicon strategic solution) BUs
have the consensus to propose AMD as the maintainer
of Platform/AMD folder. The folder and edk2 package
structure under Platform/AMD is defined in the
Readme for the further upstream works.

We still leave the existing maintainers for
AMD OverdriveBoard which is no conflicts with this
change.

Signed-off-by: Abner Chang <abner.chang@amd.com>
Cc: Leif Lindholm <quic_llindhol@quicinc.com>
Cc: Michael D Kinney <michael.d.kinney@intel.com>
Cc: Abdul Lateef Attar <abdattar@amd.com>
Cc: Eric Xing <eric.xing@amd.com>
Cc: Jiangang He <jiangang.he@amd.com>
Reviewed-by: Leif Lindholm <quic_llindhol@quicinc.com>