SbsaQemu: enlarge BL1/FIP sizes
A bug in TF-A created a discrepancy in BL1 size, and hence fip base
address. Recent changes in progress make BL1 bigger than 0x8000
which EDK2 uses.

TF-A has been fixed to change BL1 size to be 0x12000 consistently.
So let's follow it.

This is a breaking change. Binaries in edk2-non-osi will be updated
in near-lockstep.

Signed-off-by: Marcin Juszkiewicz <marcin.juszkiewicz@linaro.org>
Reviewed-by: Leif Lindholm <quic_llindhol@quicinc.com>