Platform/ARM: Fix the build failure DxeCore
The edk2 commit 7284c44951cffd52f1a08367d91b5c63c7bb9c16
introduces a new library ImagePropertiesRecordLib to
consolidate the logic for manipulating the image properties
used to track runtime images and to remove duplicate code.
This library is further linked with DxeCore.

Therefore, add an instance of ImagePropertiesRecordLib to
the ArmVExpress.dsc.inc file so that it is included for Arm
platforms.

Signed-off-by: Levi Yun <yeoreum.yun@arm.com>
Reviewed-by: Thomas Abraham <thomas.abraham@arm.com>
Reviewed-by: Sami Mujawar <sami.mujawar@arm.com>
Tested-by: Pierre Gondois <pierre.gondois@arm.com>
Tested-by: Nuno Lopes <Nuno.Lopes@arm.com>
Tested-by: Himanshu Sharma <Himanshu.Sharma@arm.com>