Vlv2TbltDevicePkg: Fix DxeCore Build Failures
Commit 7284c44 in edk2 introduces an incompatibility that causes any
project that uses DxeMain.inf to fail to build. This is due to the
addition of ImagePropertiesRecordLib, and a new added dependency on
that library in DxeMain. Platforms will not have this LibraryClass
defined in their DSC yet and hence currently fail to build.

This changes addes ImagePropertiesRecordLib to Vlv2TbltDevicePkg
and resolves the build failure.

Cc: Chasel Chiu <chasel.chiu@intel.com>
Cc: Ray Ni <ray.ni@intel.com>
Cc: Sai Chaganty <rangasai.v.chaganty@intel.com>
Cc: Liming Gao <gaoliming@byosoft.com.cn>
Reviewed-by: Michael D Kinney <michael.d.kinney@intel.com>
Cc: Zailiang Sun <zailiang.sun@intel.com>
Cc: Yi Qian <yi.qian@intel.com>
Cc: Taylor Beebe <taylor.d.beebe@gmail.com>
Reviewed-by: Michael Kubacki <michael.kubacki@microsoft.com>
Signed-off-by: Nate DeSimone <nathaniel.l.desimone@intel.com>