Platform/ARM: Juno: Generate _CPC objects for JunoR2
The SsdtCpuTopologyGenerator can generate _CPC objects.
This is done by querying the SCP for the relevant performance
state information through SCMI. CM_ARM_CPC_INFO are then populated
and used to generate _CPC objects in the Ssdt Cpu topology.

Use the DynamicTablesScmiInfoLib and add the handling to generate
_CPC information.

Note that using _CPC is only possible if SCP is correctly tuned
to advertise performance levels on an abstract and unified scale.
A basic check is done to prevent the _CPC generation otherwise.

Perf level values used for testing:
- little CPUs OPPs: [181, 322, 383] * 1000
- big CPUs OPPs: [512, 833, 1024] * 1000

Also make use of the newly added PcdDevelopmentPlatformRelaxations
and add a new ENABLE_CPC build parameter ('-D ENABLE_CPC') to enable
CPC generation.

Signed-off-by: Pierre Gondois <pierre.gondois@arm.com>
Reviewed-by: Sami Mujawar <sami.mujawar@arm.com>