Silicon/ARM/NeoverseN1Soc: Consume N1SdpNtFwConfigPei supplied data
N1SdpNtFwConfigPei PEIM extracts platform information from NT_FW_CONFIG
and provides it to other modules as a PPI and a HOB. PlatformLibMem
then consumes these values in the form of a PPI during PEI phase.
The previous approach of fetching platform information from a fixed
address is also removed in this patch.

Signed-off-by: sahil <sahil@arm.com>
Reviewed-by: Sami Mujawar <sami.mujawar@arm.com>